# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Signal Protocol API",
    "version": "1.0.0"
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "Unique identifier for the user."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration successful, returns pre-key bundle.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "preKeyBundle": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "userId is required."
          },
          "409": {
            "description": "User already registered."
          },
          "500": {
            "description": "Registration failed."
          }
        }
      }
    },
    "/prekey/{userId}": {
      "get": {
        "summary": "Retrieve pre-key bundle for a user.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "User ID to fetch pre-key bundle for."
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved pre-key bundle.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "preKeyBundle": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Failed to get PreKey bundle."
          }
        }
      }
    },
    "/session": {
      "post": {
        "summary": "Establish a session using a pre-key bundle.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "recipientId": {
                    "type": "string"
                  },
                  "preKeyBundle": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session established successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Failed to establish session."
          }
        }
      }
    },
    "/send": {
      "post": {
        "summary": "Send an encrypted message.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "senderId": {
                    "type": "string"
                  },
                  "recipientId": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message encrypted and sent successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "encryptedMessage": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters."
          },
          "404": {
            "description": "Sender not found."
          },
          "500": {
            "description": "Failed to send message."
          }
        }
      }
    },
    "/receive": {
      "post": {
        "summary": "Receive and decrypt a message.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipientId": {
                    "type": "string"
                  },
                  "senderId": {
                    "type": "string"
                  },
                  "encryptedMessage": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message decrypted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters."
          },
          "404": {
            "description": "Recipient not found."
          },
          "500": {
            "description": "Failed to decrypt message."
          }
        }
      }
    }
  },
  "components": {},
  "tags": [
    {
      "name": "Users",
      "description": "User management and protocol initialization."
    }
  ]
}